requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0      # SQLite async driver (dev/local)
asyncpg==0.29.0        # PostgreSQL async driver (production)
alembic==1.13.2        # Migrations (optional, for future use)

# Auth
pyjwt==2.9.0

# Scheduling
apscheduler==3.10.4

# Email
boto3==1.35.0
jinja2==3.1.4

# AI
httpx==0.27.2          # Async HTTP client for Claude API

# Data processing
openpyxl==3.1.5        # For Excel import migration
inflection==0.5.1      # Singularization fallback

# Utilities
python-dotenv==1.0.1
